requirements.txt
blinker==1.9.0
click==8.3.3
colorama==0.4.6
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyMySQL==1.1.2
python-dotenv==1.2.2
Werkzeug==3.1.8
